Create Current_Go
Hey Guys!
So I calculated +/- enrichments for all the tests and then took the top ten GO ids that were the most different from the control and added them all to a list. I then took the set of that list so it would only have unique variables and it came up with this list of 22 GO ids. I mapped the functions to all of the GO ids and these are the results!